model_service/requirement.txt
flask
torch
torchvision
Pillow
clearml
scikit-learn
pandas
numpy
pyarrow
fastparquet